Per-speaker font handling
