+++ public/dialogue/dialogue.yarn
@@ -1,9 +1,9 @@
 title: Start
 ---
-Eva: what's up, homosexuals?
+Eva: what’s up, homosexuals?
 Eva: have u ever thought deeply about... #emote:thinking
 Eva: [shake]veggie corndogs[/shake] #emote:flushed
-Sam: big brain scratcher #emote:unamused
-Eva: [wave]STILL EATIN' EM![/wave] #emote:determined #big
+Sam: *mrowrrr* #emote:unamused
+Eva: [wave]STILL EATIN’ EM![/wave] #emote:determined #big
 Eva: ok bye
 ===
